refactor: extract models retrieval to a separate service
